pyproject.toml
[project]
name = "lumina"
version = "0.1.0"
description = "基于 adbutils 与 OpenCV 的 FGO 自动化脚本，当前固定支持可通过 ADB 控制的 1920x1080 设备"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adbutils>=2.12.0",
    "numpy>=2.4.2",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "onnxscript>=0.6.2",
    "opencv-python>=4.13.0.92",
    "paddleocr>=2.10.0",
    "paddlepaddle==3.0.0",
    "path>=17.1.1",
    "pillow>=12.1.1",
    "pyside6>=6.11.0",
    "pyyaml>=6.0.3",
    "ultralytics>=8.4.21",
]
